Temp.

+++ OPOS.P1.Lib/Threading/CustomSchedulerSettings.cs
@@ -3,6 +3,6 @@
     public record CustomSchedulerSettings
     {
         public int MaxCores { get; init; }
-        public long MaxConcurrentTasks { get; init; }
+        public int MaxConcurrentTasks { get; init; }
     }
 }
